feat: 实现 lifecycleManagedByParent API
