+++ packages/debug/src/browser/debug-session-connection.ts
@@ -12,74 +12,11 @@ import {
 } from '@opensumi/ide-core-browser';
 import { IWebSocket } from '@opensumi/ide-connection';
 import { OutputChannel } from '@opensumi/ide-output/lib/browser/output.channel';
-import { DEBUG_REPORT_NAME, getSequenceId, IDebugSessionManager } from '../common';
+import { DebugEventTypes, DebugExitEvent, DebugRequestTypes, DEBUG_REPORT_NAME, getSequenceId, IDebugSessionManager } from '../common';
 import { DebugConfiguration } from 'vscode';
 import { DebugSessionManager } from './debug-session-manager';
 
-export interface DebugExitEvent {
-  code?: number;
-  reason?: string | Error;
-}
-
 export type DebugRequestHandler = (request: DebugProtocol.Request) => any;
-
-export interface DebugRequestTypes {
-  'attach': [DebugProtocol.AttachRequestArguments, DebugProtocol.AttachResponse];
-  'breakpointLocations': [DebugProtocol.BreakpointLocationsArguments, DebugProtocol.BreakpointLocationsResponse];
-  'completions': [DebugProtocol.CompletionsArguments, DebugProtocol.CompletionsResponse];
-  'configurationDone': [DebugProtocol.ConfigurationDoneArguments, DebugProtocol.ConfigurationDoneResponse];
-  'continue': [DebugProtocol.ContinueArguments, DebugProtocol.ContinueResponse];
-  'disconnect': [DebugProtocol.DisconnectArguments, DebugProtocol.DisconnectResponse];
-  'evaluate': [DebugProtocol.EvaluateArguments, DebugProtocol.EvaluateResponse];
-  'exceptionInfo': [DebugProtocol.ExceptionInfoArguments, DebugProtocol.ExceptionInfoResponse];
-  'goto': [DebugProtocol.GotoArguments, DebugProtocol.GotoResponse];
-  'gotoTargets': [DebugProtocol.GotoTargetsArguments, DebugProtocol.GotoTargetsResponse];
-  'initialize': [DebugProtocol.InitializeRequestArguments, DebugProtocol.InitializeResponse];
-  'launch': [DebugProtocol.LaunchRequestArguments, DebugProtocol.LaunchResponse];
-  'loadedSources': [DebugProtocol.LoadedSourcesArguments, DebugProtocol.LoadedSourcesResponse];
-  'modules': [DebugProtocol.ModulesArguments, DebugProtocol.ModulesResponse];
-  'next': [DebugProtocol.NextArguments, DebugProtocol.NextResponse];
-  'pause': [DebugProtocol.PauseArguments, DebugProtocol.PauseResponse];
-  'restart': [DebugProtocol.RestartArguments, DebugProtocol.RestartResponse];
-  'restartFrame': [DebugProtocol.RestartFrameArguments, DebugProtocol.RestartFrameResponse];
-  'reverseContinue': [DebugProtocol.ReverseContinueArguments, DebugProtocol.ReverseContinueResponse];
-  'scopes': [DebugProtocol.ScopesArguments, DebugProtocol.ScopesResponse];
-  'setBreakpoints': [DebugProtocol.SetBreakpointsArguments, DebugProtocol.SetBreakpointsResponse];
-  'setExceptionBreakpoints': [DebugProtocol.SetExceptionBreakpointsArguments, DebugProtocol.SetExceptionBreakpointsResponse];
-  'setExpression': [DebugProtocol.SetExpressionArguments, DebugProtocol.SetExpressionResponse];
-  'setFunctionBreakpoints': [DebugProtocol.SetFunctionBreakpointsArguments, DebugProtocol.SetFunctionBreakpointsResponse];
-  'setVariable': [DebugProtocol.SetVariableArguments, DebugProtocol.SetVariableResponse];
-  'source': [DebugProtocol.SourceArguments, DebugProtocol.SourceResponse];
-  'stackTrace': [DebugProtocol.StackTraceArguments, DebugProtocol.StackTraceResponse];
-  'stepBack': [DebugProtocol.StepBackArguments, DebugProtocol.StepBackResponse];
-  'stepIn': [DebugProtocol.StepInArguments, DebugProtocol.StepInResponse];
-  'stepInTargets': [DebugProtocol.StepInTargetsArguments, DebugProtocol.StepInTargetsResponse];
-  'stepOut': [DebugProtocol.StepOutArguments, DebugProtocol.StepOutResponse];
-  'terminate': [DebugProtocol.TerminateArguments, DebugProtocol.TerminateResponse];
-  'terminateThreads': [DebugProtocol.TerminateThreadsArguments, DebugProtocol.TerminateThreadsResponse];
-  'threads': [DebugProtocol.ThreadsArguments | null, DebugProtocol.ThreadsResponse];
-  'variables': [DebugProtocol.VariablesArguments, DebugProtocol.VariablesResponse];
-  'cancel': [DebugProtocol.CancelArguments, DebugProtocol.CancelResponse];
-}
-
-export interface DebugEventTypes {
-  'breakpoint': DebugProtocol.BreakpointEvent;
-  'capabilities': DebugProtocol.CapabilitiesEvent;
-  'continued': DebugProtocol.ContinuedEvent;
-  'exited': DebugExitEvent;
-  'initialized': DebugProtocol.InitializedEvent;
-  'loadedSource': DebugProtocol.LoadedSourceEvent;
-  'module': DebugProtocol.ModuleEvent;
-  'output': DebugProtocol.OutputEvent;
-  'process': DebugProtocol.ProcessEvent;
-  'stopped': DebugProtocol.StoppedEvent;
-  'terminated': DebugProtocol.TerminatedEvent;
-  'thread': DebugProtocol.ThreadEvent;
-  'progressStart': DebugProtocol.ProgressStartEvent;
-  'progressUpdate': DebugProtocol.ProgressUpdateEvent;
-  'progressEnd': DebugProtocol.ProgressEndEvent;
-  'invalidated': DebugProtocol.InvalidatedEvent;
-}
 
 const standardDebugEvents = new Set<Required<keyof DebugEventTypes>>([
   'breakpoint',
