+++ packages/extension/src/common/vscode/debug.ts
@@ -9,6 +9,7 @@ export type DebugSessionUUID = string;
 
 export interface IStartDebuggingOptions {
   parentSessionID?: DebugSessionUUID;
+  lifecycleManagedByParent?: boolean;
   repl?: 'separate' | 'mergeWithParent';
   noDebug?: boolean;
   compact?: boolean;
